refactor: Entity 어노테이션 변경 및 employeeNumber, date 필드 초기화 적용

+++ src/main/java/com/sprint/hrbank/repository/DepartmentRepository.java
@@ -3,6 +3,4 @@ package com.sprint.hrbank.repository;
 import com.sprint.hrbank.entity.Department;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-interface DepartmentRepository extends JpaRepository<Department, Long> {
-
-}
+interface DepartmentRepository extends JpaRepository<Department, Integer> {}
